# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gosso - SSO Authentication Server
  description: |
    A production-ready Single Sign-On (SSO) authentication server built with Go.

    Supports:
    - Username/password login with MFA (TOTP + backup codes)
    - OAuth 2.0 Authorization Server (authorization code, refresh token, client credentials)
    - OpenID Connect Provider (ID token, discovery, JWKS, userinfo)
    - Social login (Google, GitHub, WeChat)
    - Admin account management
    - Email/phone verification and password reset
  version: 1.0.0
  contact:
    name: rushairer
    url: https://github.com/rushairer/gosso
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/oauth2/clients:
    get:
      tags: [OAuth2 Clients]
      summary: List OAuth2 clients
      description: Returns all OAuth2 clients owned by the authenticated account.
      operationId: listClients
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of OAuth2 clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OAuth2Client"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [OAuth2 Clients]
      summary: Register a new OAuth2 client
      description: Creates a new OAuth2 client for the authenticated account. Returns the client ID and secret.
      operationId: registerClient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/RegisterClientResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/oauth2/clients/{client_id}:
    parameters:
      - $ref: "#/components/parameters/ClientID"

    get:
      tags: [OAuth2 Clients]
      summary: Get OAuth2 client details
      operationId: getClient
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/OAuth2Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [OAuth2 Clients]
      summary: Update OAuth2 client
      operationId: updateClient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/OAuth2Client"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [OAuth2 Clients]
      summary: Delete OAuth2 client
      operationId: deleteClient
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Client deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                code: 200
                message: success
                data: client deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # OAuth2 Protocol
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login, token, or social login endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth with client_id as username and client_secret as password
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection via double-submit cookie pattern.
        1. GET any endpoint to receive `csrf_token` cookie
        2. Read cookie value and send as `X-CSRF-Token` header on POST/PUT/DELETE
        3. Not required when using `Authorization: Bearer` header (JWT auth)

  # ── Parameters ──
  parameters:
...
    ClientID:
      name: client_id
      in: path
      required: true
      schema:
        type: string
      description: OAuth2 client ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            message: invalid request body
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            message: unauthorized
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 404
            message: not found
...
  schemas:
...
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          description: Response data (type varies by endpoint)
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: error message
...
    OAuth2Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        client_id:
          type: string
        name:
          type: string
        description:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        grant_types:
          type: array
          items:
            type: string
            enum: [authorization_code, refresh_token, client_credentials]
        scopes:
          type: array
          items:
            type: string
        is_confidential:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RegisterClientRequest:
      type: object
      required: [name, redirect_uris]
      properties:
        name:
          type: string
          example: My App
        description:
          type: string
          example: My OAuth2 application
        redirect_uris:
          type: array
          minItems: 1
          items:
            type: string
            format: uri
          example: ["https://app.example.com/callback"]
        grant_types:
          type: array
          items:
            type: string
            enum: [authorization_code, refresh_token, client_credentials]
          default: [authorization_code]
        scopes:
          type: array
          items:
            type: string
          example: ["openid", "profile", "email"]
        is_confidential:
          type: boolean
          default: false
    RegisterClientResponse:
      type: object
      properties:
        client_id:
          type: string
          description: Unique client identifier
        client_secret:
          type: string
          description: Client secret (only shown once, only for confidential clients)
        name:
          type: string
    UpdateClientRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        grant_types:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string

    # -- OAuth2 Protocol --
...